requirements.txt
flask
openai
python-dotenv
requestsflask-cors
